Spliting functions into different files
