+++ football-annoucer/settings.json
@@ -1,9 +1,20 @@
 {
-    "daily_cron_job": "0 12 * * *",
+    "daily_cron_job": "*/2 * * * *",
     "bot_prefix_command": "+",
     "football_data_api": "http://api.football-data.org/v2/",
-    "football_data_competitions": "2017",
+    "football_data_competitions": [
+        {
+            "name": "Liga NOS",
+            "id": 2017,
+            "active": false
+        },
+        {
+            "name": "Euro 2020",
+            "id": 2018,
+            "active": true
+        }
+    ],
     "is_daily_job_active": true,
-    "announcements_channel": "",
+    "announcements_channel": "719536540791078966",
     "persons_in_charge": ""
 }
